+++ vasttools/query.py
@@ -419,12 +419,11 @@ class Query:
             self.logger.critical("Only Stokes I are supported with tiles!")
             return False
 
-        if self.settings['tiles'] and not self.settings['no_rms']:
-            self.logger.warning(
-                "RMS measurements are not supported with tiles!"
+        if self.settings['tiles'] and self.settings['islands']:
+            self.logger.critical(
+                "Only component catalogues are supported with tiles!"
             )
-            self.logger.warning("Turning RMS measurements off.")
-            self.settings['no_rms'] = True
+            return False
 
         if self.settings['islands']:
             self.logger.warning(
@@ -1775,17 +1774,19 @@ class Query:
             dir_name = "TILES"
 
             image_file_fmt = (
-                "image.i.{}.SB{}.cont"
+                "image.{}.{}.SB{}.cont"
                 ".taylor.0.restored.fits".format(
-                    row.field, row.sbid
+                    self.settings['stokes'].lower(), row.field, row.sbid
                 )
             )
+
             if self.corrected_data:
                 img_dir += "_CORRECTED"
                 rms_dir += "_CORRECTED"
                 image_file_fmt = image_file_fmt.replace(".fits",
                                                         ".corrected.fits"
                                                         )
+            rms_file_fmt = f"noiseMap.{image_file_fmt}"
 
         else:
             dir_name = "COMBINED"
@@ -1812,16 +1813,13 @@ class Query:
             image_file_fmt
         )
 
-        if self.settings['tiles']:
-            rms_file = "N/A"
-        else:
-            rms_file = os.path.join(
-                self.base_folder,
-                epoch_string,
-                dir_name,
-                rms_dir,
-                rms_file_fmt
-            )
+        rms_file = os.path.join(
+            self.base_folder,
+            epoch_string,
+            dir_name,
+            rms_dir,
+            rms_file_fmt
+        )
 
         return selavy_file, image_file, rms_file